+++ src/components/Player.vue
@@ -260,7 +260,6 @@
       >
         <span
           class="icon"
-          v-if="reminder.role !== 'custom'"
           :style="{
             backgroundImage: `url(${
               reminder.image && grimoire.isImageOptIn
@@ -1072,7 +1071,7 @@ li.move:not(.from) .player .overlay svg.move {
     }
   }
 
-  &.custom {
+  &.fabled {
     .icon {
       display: none;
     }